+++ src/main/java/br/com/challenge/api/service/presentation/dto/ValorMedioVendaECompra.java
@@ -1,7 +1,5 @@
 package br.com.challenge.api.service.presentation.dto;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -15,8 +13,8 @@ import lombok.Setter;
 @Getter
 public class ValorMedioVendaECompra {
 
-	  private BigDecimal valorMedioCompra;	
-	  private BigDecimal valorMedioVenda;	
+	  private Double valorMedioCompra;	
+	  private Double valorMedioVenda;	
 	 
 	   
 }
